add abstract model. add person model and edit genre model with abstract
