+++ fastpapi-solution/src/models/person.py
@@ -1,0 +1,12 @@
+import uuid
+from typing import ClassVar
+
+from models.base import BaseMovie
+
+
+class Person(BaseMovie):
+    index: ClassVar[str] = 'persons'
+
+    full_name: str
+    role: str
+    film_ids: list[uuid.UUID]
